requirements.txt
Pillow>=10.3
numpy>=1.26
streamlit>=1.34
